move options in separate file
